user-service/pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.128.0",
    "pydantic[email]>=2.12.5",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
